convert to double again

+++ src/exp_ms.py
@@ -188,11 +188,11 @@ def run_ms(
 def append_prev_info(x_it_, scalar_list):
     x_it_.type(torch.DoubleTensor)
     for s in scalar_list:
-        s.double()
+        s.type(torch.DoubleTensor)
         print("stype: ", s.type())
         print("x_it_ type: ", x_it_.type())
         x_it_ = torch.cat(
-            [x_it_, s.type(torch.FloatTensor).view(tensor_length(s))]
+            [x_it_, s.type(torch.DoubleTensor).view(tensor_length(s))]
         )
     return x_it_
